Update migration to ensure budget costs are floats

+++ src/main/resources/db/migration/V157__Add_Budget_View_Tables.sql
@@ -1,9 +1,9 @@
 CREATE TABLE `Budgets` (
   `Id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   `ScheduleId` int(11) NOT NULL,
-  `TaCost` int(11) NULL,
-  `ReaderCost` int(11) NULL,
-  `LecturerCost` int(11) NULL
+  `TaCost` FLOAT NULL,
+  `ReaderCost` FLOAT NULL,
+  `LecturerCost` FLOAT NULL
 );
 
 CREATE TABLE `BudgetScenarios` (
@@ -21,8 +21,8 @@ CREATE TABLE `SectionGroupCosts` (
   `CourseNumber` varchar(7) NOT NULL,
   `SequencePattern` varchar(7) NOT NULL,
   `TermCode` VARCHAR(7) NOT NULL,
-  `UnitsLow` float DEFAULT NULL,
-  `UnitsHigh` float DEFAULT NULL,
+  `UnitsLow` FLOAT DEFAULT NULL,
+  `UnitsHigh` FLOAT DEFAULT NULL,
   `EffectiveTermCode` varchar(6) NOT NULL,
   `InstructorId` int(11) NULL,
   `OriginalInstructorId` int(11) NULL,
@@ -30,7 +30,7 @@ CREATE TABLE `SectionGroupCosts` (
   `ReaderCount` int(11) NULL,
   `SectionCount` int(11) NULL,
   `Enrollment` int(11) NULL,
-  `InstructorCost` int(11) NULL,
+  `InstructorCost` FLOAT NULL,
   `ReplacementReason` varchar(50) NULL
 );
 
@@ -69,6 +69,6 @@ CREATE TABLE `InstructorCosts` (
   `Id` INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
   `BudgetId` int(11) NOT NULL,
   `InstructorId` int(11) NOT NULL,
-  `Cost` int(11) NULL,
+  `Cost` FLOAT NULL,
   `Lecturer` int(11) DEFAULT 0 NOT NULL
 );
